+++ datastore/couchDB/utilities.go
@@ -12,6 +12,7 @@ import (
 	"github.com/accedian/adh-gather/logger"
 	"github.com/accedian/adh-gather/models"
 	tenmod "github.com/accedian/adh-gather/models/tenant"
+	setops "github.com/adam-hanna/arrayOperations"
 	couchdb "github.com/leesper/couchdb-golang"
 )
 
@@ -737,4 +738,14 @@ func generatePaginationQueryParams(startKey string, limit int64, includeDocs boo
 	params.Add(descendingQueryParamStr, strconv.FormatBool(descending))
 
 	return params
+}
+
+// Return the string union between the provided arrays
+func listUnion(listA []string, listB []string) []string {
+	return setops.UnionString(listA, listB)
+}
+
+// return the string intersection between the provided arrays
+func listIntersection(listA []string, listB []string) []string {
+	return setops.IntersectString(listA, listB)
 }